Refactor import statements in models
