+++ src/img2img/models/pix2pix/train.py
@@ -8,9 +8,13 @@ from tqdm import tqdm
 
 from img2img import cfg
 from img2img.data import create_dataset
-from img2img.pix2pix.utils import save_checkpoint, load_checkpoint, save_some_examples
-from img2img.pix2pix.generator import Generator
-from img2img.pix2pix.discriminator import Discriminator
+from img2img.models.pix2pix.utils import (
+    save_checkpoint,
+    load_checkpoint,
+    save_some_examples,
+)
+from img2img.models.pix2pix.generator import Generator
+from img2img.models.pix2pix.discriminator import Discriminator
 
 _WRITER = SummaryWriter("runs/expirement_1")
