+++ src/img2img/models/tunit/train.py
@@ -12,11 +12,15 @@ from typing import Dict
 
 from img2img import cfg
 from img2img.data import create_dataset
-from img2img.tunit.utils import save_checkpoint, load_checkpoint, save_some_examples
-from img2img.tunit.generator import Generator
-from img2img.tunit.discriminator import Discriminator
-from img2img.tunit.guiding_network import GuidingNetwork
-from img2img.tunit.utils import (
+from img2img.models.tunit.utils import (
+    save_checkpoint,
+    load_checkpoint,
+    save_some_examples,
+)
+from img2img.models.tunit.generator import Generator
+from img2img.models.tunit.discriminator import Discriminator
+from img2img.models.tunit.guiding_network import GuidingNetwork
+from img2img.models.tunit.utils import (
     calc_contrastive_loss,
     calc_iic_loss,
     calc_adv_loss,
