+++ src/img2img/models/unit/utils.py
@@ -3,12 +3,12 @@ import math
 from pathlib import Path
 
 # import yaml
-import time
 
 import torch
 from torch import nn
 from torch.utils.data import DataLoader
-from torch.utils.tensorboard import SummaryWriter
+
+# from torch.utils.tensorboard import SummaryWriter
 from torch.nn import init
 from torch.optim import lr_scheduler
 from torchvision.utils import save_image
